LayoutTests/platform/blackberry/accessibility/placeholder-expected.txt


This test makes sure that the placeholder is returned as the correct attribute

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL fieldElement.stringAttributeValue('AXPlaceholderValue') should be search. Was .
PASS successfullyParsed is true

TEST COMPLETE
